Update ULI_CheckDigitCalc_wLEI_MultiLoanNums.jScript
Corrects for erroneous validation of check digits 99, 00, and 01. These are always invalid check digits. See for example that source 065 returns a check digit of 97 not 00. Here, 06597 was returning valid but 06500 was also erroneously returning valid. Similarly, 097 returns check digit of 98 where 09798 is valid but 09701 was erroneously returning valid. Finally, 129 returns check digit of 02 where 12902 is valid but 12999 was erroneously returning valid.